fix: resolve issues found during end-to-end verification
- API used Identity cookie scheme instead of JWT bearer, causing 401 on all protected endpoints (switch to AddIdentityCore + explicit scheme)
- Enums serialized as numbers; now serialized as strings to match frontend and Postman (product size kept as integer)
- New cart items failed to insert: EF treated client-generated Guid keys as store-generated (configure ValueGeneratedNever)
- Cart handlers misused DbSet.Update() on EF-tracked entities
- Checkout built order item names after clearing the cart
- Dev API port aligned to 5000 (frontend, Postman and README expect it)

+++ backend/ECommerce.Application/Products/Commands/ProductCommands.cs
@@ -33,7 +33,7 @@ public class CreateProductCommandHandler : IRequestHandler<CreateProductCommand,
         await _uow.SaveChangesAsync(ct);
 
         return Result<ProductDto>.Success(new ProductDto(product.Id, product.Code, product.Name,
-            product.Description, product.ImageUrl, product.Size, product.Color, product.Price, product.Stock, product.Stock > 0));
+            product.Description, product.ImageUrl, (int)product.Size, product.Color, product.Price, product.Stock, product.Stock > 0));
     }
 }
 
@@ -53,7 +53,7 @@ public class UpdateProductCommandHandler : IRequestHandler<UpdateProductCommand,
         await _uow.SaveChangesAsync(ct);
 
         return Result<ProductDto>.Success(new ProductDto(product.Id, product.Code, product.Name,
-            product.Description, product.ImageUrl, product.Size, product.Color, product.Price, product.Stock, product.Stock > 0));
+            product.Description, product.ImageUrl, (int)product.Size, product.Color, product.Price, product.Stock, product.Stock > 0));
     }
 }
